requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
openai-whisper==20231117
deep-translator==1.11.4
transformers==4.45.0
torch>=2.2.0
yt-dlp==2024.12.23
python-dotenv==1.0.0
validators==0.22.0
requests==2.31.0
moviepy==1.0.3
networkx==3.2
sentencepiece==0.2.0
soundfile==0.12.1